deleted data_manager, comment_data_manager created, timegenerator --> util
