+++ util.py
@@ -1,0 +1,6 @@
+import datetime
+
+
+def submission_time_generator():
+    submission_time = datetime.datetime.now()
+    return submission_time
